Reworked Login Workflow so that the error styles are correctly applied and removed in the UI.
Additionally improved the Validation States for the Login Workflow. A new State has been added that checks if the email is within a threshold of 3 in Levenstheins Distance. If so the E-Mail is corrected via the Deep Seek API. If the email is further away from a know e-mail address it is no longer autocorrected but a new error message is displayed.